+++ src/index.ts
@@ -4,10 +4,14 @@ import instanceConfigFields from './instanceConfigFields';
 import { activeDirectorySteps } from './steps/active-directory';
 import { intuneSteps } from './steps/intune';
 import { IntegrationConfig } from './types';
-import validateInvocation from './validateInvocation';
+import { getStepStartStates } from './utils/getStepStartStates';
+import { validateInvocation } from './validateInvocation';
+
+export const integrationSteps = [...activeDirectorySteps, ...intuneSteps];
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields,
   validateInvocation,
-  integrationSteps: [...activeDirectorySteps, ...intuneSteps],
+  getStepStartStates,
+  integrationSteps,
 };