+++ src/steps/active-directory/__tests__/converters.test.ts
@@ -20,22 +20,19 @@ import exampleOrganization from './fixtures/exampleOrganization';
 
 describe('createAccountEntityWithOrganization', () => {
   test('properties transferred', () => {
+    const exampleIntuneConfig = {
+      mobileDeviceManagementAuthority: 'mobileDeviceManagementAuthority',
+      subscriptionState: 'subscriptionState',
+      intuneAccountID: 'intuneAccountID',
+    };
     const accountEntity = createAccountEntityWithOrganization(
       exampleIntegrationInstance,
       exampleOrganization,
+      exampleIntuneConfig,
     );
-    expect(accountEntity).toEqual({
-      _class: ['Account'],
-      _key: 'microsoft_365_account-the-instance-id',
-      _type: 'microsoft_365_account',
-      _rawData: [{ name: 'default', rawData: exampleOrganization }],
-      name: 'Default Directory',
-      id: '1111111-1111-1111-1111-111111111111',
-      displayName: 'instance.config.name configured by customer',
-      defaultDomain: 'verifiedDomain.onmicrosoft.com',
-      organizationName: 'Default Directory',
-      verifiedDomains: ['verifiedDomain.onmicrosoft.com'],
-    });
+    expect(accountEntity).toMatchSnapshot(
+      'createAccountEntityWithOrganization',
+    );
   });
 });